move indexer constants to indexer-consts.ts
move BlockPayload to model

move formatEventId to one of the model classes
