+++ substrate-query-framework/index-builder/src/indexer/index.ts
@@ -3,4 +3,4 @@ export * from './BlockProducer';
 export * from './IndexBuilder';
 export * from './PooledExecutor';
 export * from './IndexerStatusService';
-export * from './redis-consts';
+export * from './redis-keys';
