+++ substrate-query-framework/index-builder/src/indexer/redis-keys.ts
@@ -1,6 +1,5 @@
 export const INDEXER_NEW_HEAD_CHANNEL = 'hydra:indexer:head:new';
 export const INDEXER_HEAD_BLOCK = 'hydra:indexer:head';
-export const INDEXER_RECENTLY_COMPLETE_BLOCKS = 'hydra:indexer:recent:blocks';
 export const BLOCK_COMPLETE_CHANNEL = 'hydra:indexer:complete:block';
 export const BLOCK_START_CHANNEL = 'hydra:indexer:started:block';
 export const EVENT_LAST = 'hydra:indexer:last:event';
