+++ substrate-query-framework/index-builder/src/model/index.ts
@@ -2,4 +2,5 @@ export * from './QueryEvent';
 export * from './QueryEventBlock';
 export * from './QueryEventProcessingPack';
 export * from './substrate-interfaces';
+export * from './BlockPayload';
